+++ kaseev.andrey/S1/AdditionalFunctions.hpp
@@ -5,7 +5,7 @@
 namespace kaseev
 {
   int readList(const std::string &line, kaseev::List<std::pair<std::string, kaseev::List<unsigned long long>>>& arr, bool &marker);
-  int sumNumbersInString(const std::string& line);
+  unsigned long long sumNumbersInString(const std::string& line);
   void sumNumbersInArray(const kaseev::List<std::pair<std::string, kaseev::List<int>>>& sums, bool &marker);
   void printListNames(const kaseev::List<std::pair<std::string, kaseev::List<unsigned long long>>>& arr);
   kaseev::List<std::pair<std::string, kaseev::List<int>>>
